[FIX] Cast argmax/argmin concretize result to int32
np.argmax/np.argmin return int64 on 64-bit machines, but the node
declares tl.int32 as its dtype. This mismatch causes TensorHandle
validation to fail. Cast the result to int32 before wrapping.

+++ triton_viz/clients/symbolic_engine.py
@@ -1224,7 +1224,7 @@ class ReduceSymbolicExpr(SymbolicExpr):
             else bool(self.keepdims)
         )
         if self.op in ("argmax", "argmin"):
-            result = np_op(input_concrete.data, axis=axis_val)
+            result = np_op(input_concrete.data, axis=axis_val).astype(np.int32)
             if keepdims_val:
                 result = np.expand_dims(
                     result, axis=axis_val if axis_val is not None else 0
